yt-video/requirements.txt
# Video İşleme
moviepy>=2.0.0.dev2
numpy>=1.24.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# HTTP ve Cloud
requests>=2.31.0
boto3>=1.34.0
fal-client>=0.4.0

# Watch (Geliştirme için)
watchfiles>=0.21.0
